removing unnecessary functions and consolidating modules
